requirements.txt
setuptools==69.0.3
selenium==4.17.2
allure-behave==2.13.2
behave==1.2.6
requests==2.31.0